added 'Enroll User' action to the message campaign channel
https://trello.com/c/MxXuzgod

Added the 'Enroll user' event as an channel action in the
tasks channel definition in the message campaign module.

Because of the defined types for action fields in
the Tasks module, the enroll user handler have to check
type of event parameters (reference date, reference time,
start time) and converts them to the correct type
required by CampaignRequest.

Change-Id: I4329a22a7546f002860bc3a952073b821eaf0875