BUG: IVR Logging: Change date search to what is outlined in spec
https://trello.com/c/BWxXnCjw/

Changed date search in IVR Call log from being a single date range
to being 3 date ranges: one for start date, one for answer date, and
one for end date. Also added a column for answer date in the call log,
and added a seconds slider to the date selector.

Change-Id: Idd9f5b9edfb8922854d79eec56e2d7a0c4e11da6