Handle config changes via UI
https://trello.com/c/R4XrYTJA

This commit creates a skeleton approach on persisting
module properties in the database, when config source is
set to UI.
Persisting and retrieving properties has been implemented
in the new ConfigurationService class. When ConfiSource is
set to UI non-default properties are stored in the database.
Method responsible for retrieving properties merges stored
properties with default bundle properties and returns the result.
When ConfigSource is set to File, everything should work pretty
much like before these changes, however some refactoring has been done.
Some of the methods from the PlatformSettingsService have been
moved to the new ConfigurationService and merged with the new
methods.

Change-Id: I5483c9a3efd897eba04140eed4bb0f6fe2833058