Auto generate module header
https://trello.com/c/aFzHsV87

This commit modify the way in which module header is created. In past the
Header class loads entries from header.html file which was inside bundle.
Currently, this class searches css and js files in webapp/css, webapp/js
and webapp/lib folders and creates appropriate header for the module.

Change-Id: Ie425c64428ae915e2e4b697d9ac5161a99efaae5