Moved bootstrap UI to the war and separated it from OSGi startup.
https://trello.com/c/w1Ufd8v1

This commit moves the bootstrap page to the war application.
The bootstrapping is done from a separate servlet and only after its
finished the OSGi framework will start up. Unfortunately this commit called
for some duplication of javascript/css libraries. Some better way of sharing
these can be done later, as it is just jquery+boostrap and some css.
Changes to the Motech UI should not really require drastic changes for
the bootstrap view. The bootstrap page is still a jsp, as it was such
before this work. Porting it to AngularJs should be done within a separate
card.

In general, this a quicker fix for the issue of the bootstrap UI not
being usable. It does not require changes across modules using CouchDb
and as such is less intrusive and sufficient for now.

This also required a change to our startup procedure. Now, when the
bundleContext is set in the servlet context in initialization, the
HttpService is getting registered faster. Because of this a bug has
surfaced: web-security would attempt to initialize before CouchDbManager
was available. To fix this, the OsgiFrameworkService will now start the
http-bridge bundle only after CouchDbManager becomes available.
Thanks to this web-security can still start the way it did, immediately
after the http service becomes available.

Change-Id: I4b5096bc2261f44e3c1acfe3237b2480dc1fa3b9