SM: Provide basic classes for given entity
https://trello.com/c/I6iktDb3

This commit provides a first version of mechanism to create repository,
service interface and service implementation of this interface for the
given entity schema.

These 3 classes are created when the entity schema class is first time
loaded by PersistanceClassLoader. They are created and added to
classpath only if they are not present in classpath.

Change-Id: I0c4eb0ceb2289311398dbc541e48c4b1721809ca