SAP: Processing @Lookup annotation
https://trello.com/c/c8OFR9aa

This commit adds the processing of @Lookup annotations. The lookup
annotation is applied to a public method that acts as the method
for retrieving Entities from the database. The Lookup processor
determines the entity looked for by the return type of the lookup
method. It also determines the single/multiple object return if
the returned value is a collection. The lookup name is generated
either by providing the name together with the annotation or based on
the method name (converted to more user-friendly string).
The fields connected with the lookup are determined via the
method parameters. If no field is annotated with the @LookupField
annotation, all of the method paramters are considered as lookup fields.
If the paramaters are annotated, it will only pick the names from
the annotation.

Change-Id: Ic01bbd536e5aed66e40aea1feb1cc6db29752ebd