AP: @UIDisplayable annotation
https://trello.com/c/RCehJxYd

Added mechanism to finding fields or methods with @UIDisplayable
annotation inside class with @Entity annotation.

@UIDisplayable
(position=)

The @UIDisplayable allows the developer to mark a field as being in
the default display for a listing of objects. If this annotation is
present then only fields with the annotation will be displayed. If the
field is not present all fields will be displayed the ordering is
undefined, but most likely in the order fields are defined in the
object.

Change-Id: Idc9fc8860571c38f6c121404b4d60d41a9602a1c