MOTECH-901: Updated startup process to work well with tests
Code responsible for controlling the startup process(except the boostrap
page) was moved to a bundle, called osgi-platform. Thanks to this, the
same startup logic can be reused in OSGi tests, it also makes the code
more portable.

A change was made to MDS as well, instead of  wait for spring contexts
being published, MDS acts now as a bundle listner, processing bundles when
they get installed.

Another problem resolved by this commit is the fact that refreshes coming
from MDS would not work well with JUnit testing. Because of this, the
native Pax Exam container was extended, our extension starts the probe
after the tested bundle starts.

Also some code has been cleaned up and some logging has been added.

This commit will require adjustments to test code, since modules will not
be started, unless the platform-server-bundle is started in the test
framework, or a startup modules event is faked.

Change-Id: Icd86ce458ff541b0f73344699a563bd65b4acae9