Fixed daylight savings issue when scheduling with offsets
Instead of adding seconds for whatever reason, we now just add
the period itself. This is aware of daylight saving changes and it
will ensure the message will be delivered at the given time.
This issue was caught by the unit test at my local machine, since
the daylight savings time change is now within a month.

Change-Id: I4959c3b2bb3dbf4a2c96beaef6bf0eb02552ff71