MOTECH-1663 Fixed occasional problems with Commcare schema on PSQL
The Commcare module would fail to store the Commcare schema, in
case the schema contained special characters - eg. unusal letters,
not present in the english alphabet and the selected datastore was
PostgreSQL. This didn't seem to be a problem on our side, since
the String with schema is constructed properly and the same works
with MySQL datastore. As a solution, we now always escape all special
characters, before storing schema and unescape them during retrieval
of the schema from MDS. I additionaly enhanced logging, and removed
debug level log for Commcare tests. Moreover, the Commcare module
will start up properly now, even if the schema cannot be deserialized,
allowing uses to force fetch of fresh schema from CommCareHQ.

Change-Id: I24724d6dd551c1347582717b486872e7800c5d55