MOTECH-2076: CommCare module task triggers depend on xmlns
Previously the CommCare module relied on the form name to determine when to
trigger a task. Now it relies on xmlns. The event subject will look like the
following :
org.motechproject.commcare.api.forms + configName + formXMLNS

For the end user nothing changes, the display names for task triggers are the
same and depends on form name.