MOTECH-2161: Adds getEncounterByUuid lookup
This PR adds next lookup to a task data source in OpenMRS module. I had to
adjust OpenMRS model in our Motech implementation. The following changes
had to be done:

- add display field to Encounter and OpenMRSEncounter
- add display field to Location and OpenMRSFacility

OpenMRSEncounter and OpenMRSFacility are Motech representation of Encounter
and Location model.

After this changes we have getEncounterByUuid lookup in OpenMRS task data source,
with the following fields :

Encounter UUID
Encounter display
Encounter date
Provider UUID
Provider display
Location UUID
Location display
Patient UUID
Patient display

In the spec for this issue there is information about OpenMRS Form, which is
related with Encounter. However, now we don't have Motech representation of
OpenMRS Form and that's why this field is not included in Encounter fields.
I will do it in another PR.

Lookups getPatientByUuid and getPatientByMotechId will be added in the separate
PR.